// frontend/src/app/core/api/generated/model/page-response-warehouse-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WarehouseResponse } from './warehouse-response';


export interface PageResponseWarehouseResponse { 
    content?: Array<WarehouseResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    first?: boolean;
    last?: boolean;
}